Build every stop reference through one helper
trips-for-location and trips-for-route each grew their own version of
"label a stop reference with the combined ID the entry referred to it
by", and the copies had already diverged: trips-for-location dropped a
stop whose routes did not resolve, while trips-for-route kept it with an
empty route list, and only trips-for-location derived a direction from
shape geometry.

Fold both into stopReferences in reference_utils.go, keeping the
behaviour that is right in each case: routeless stops still get a
reference, so every stop ID an entry emits resolves in the block, and
both endpoints now compute direction through the DirectionCalculator,
falling back to UNKNOWN when no shape supports one.

The empty-referredID branch goes with them. Both call sites build their
stop list from the keys of the very map they pass in, so the ID is
always present and the fallback was unreachable.
